% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{autoplot.mirvie_cohort_outliers}
\alias{autoplot.mirvie_cohort_outliers}
\title{Plot a \code{mirvie_cohort_outliers} object.}
\usage{
\method{autoplot}{mirvie_cohort_outliers}(object, pcx = 1, pcy = 2, plotly = interactive(), ...)
}
\arguments{
\item{object}{A \code{mirvie_cohort_outliers} object.}

\item{pcx}{An integer between 1 and 5. The principal component that will be
on the x axis.}

\item{pcy}{An integer between 1 and 5. The principal component that will be
on the y axis.}

\item{plotly}{A flag. Make the plot interactive (with mirvie ID tooltips)?}

\item{...}{Not currently used.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or a \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.
}
\description{
A \code{mirvie_cohort_outliers} object is the output of a call to
\code{\link[=get_cohort_outliers]{get_cohort_outliers()}}.
}
