% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{check_sample_descs}
\alias{check_sample_descs}
\title{Check the sample descriptions data frame for an R&D experiment.}
\usage{
check_sample_descs(sample_descs)
}
\arguments{
\item{sample_descs}{A data frame of sample descriptions that includes
columns called \code{sample_name} and \code{condition}.}
}
\value{
\code{TRUE} (invisibly) if the check passes. Otherwise, an error is
thrown.
}
\description{
\itemize{
\item It must contain a column called \code{sample_name} with unique values.
\item It must contain a column called \code{condition}.
\item If it contains a column called \code{id_tech_rep}, then for a given
\code{id_tech_rep}, the \code{sample_name}s must have a common substring at least two
characters long.
}
}
\seealso{
Other Arkady: 
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
