% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{collapse_tech_reps}
\alias{collapse_tech_reps}
\title{Collapse (sum) technical replicates in a set of R&D experiments.}
\usage{
collapse_tech_reps(prepped_rd_input)
}
\arguments{
\item{prepped_rd_input}{A data frame. The output of a call to
\code{\link[=prep_rd_input]{prep_rd_input()}}.}
}
\value{
A dataframe with columns \code{id_genes_counts}, \code{condition} and the gene
columns.
}
\description{
Technical replicates, identified by the \code{id_tech_rep} column, are summed into
a single sample. The resulting \code{id_genes_counts} is the longest common
substring of the \code{id_genes_counts} of the input technical replicates.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
