% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather-csv.R
\name{convert_feather_dir_to_csvs}
\alias{convert_feather_dir_to_csvs}
\title{Make a directory of CSVs from a directory of feathers.}
\usage{
convert_feather_dir_to_csvs(feather_dir_path, new_dir_name = "feather-csvs")
}
\arguments{
\item{feather_dir_path}{The path to the directory containing the feathers.}

\item{new_dir_name}{The name of the new directory. This should \emph{not} be an
absolute path and rather just a name; I.e. it should not contain '/'. The
created directory will be a sibling of the one at \code{feather_dir_path}.}
}
\value{
The path to the output directory, invisibly.
}
\description{
Take a directory containing feather files and create a sibling directory with
corresponding CSVs.
}
