% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{convert_gene_names}
\alias{convert_gene_names}
\title{Convert gene names to/from ensembl.}
\usage{
convert_gene_names(x, ensembl = "from")
}
\arguments{
\item{x}{A character vector.}

\item{ensembl}{A string. Either \code{"from"} or \code{"to"}. With \code{"to"} the result is
Ensembl gene names. With \code{"from"} the result is colloquial gene names.}
}
\value{
A character vector.
}
\description{
This function works in a particular way: inputs that don't look like a gene
at all are returned as is.
}
