% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{deseq_contrasts}
\alias{deseq_contrasts}
\title{Obtain pairwise contrasts with shrunken log2FC from a DESeq  object.}
\usage{
deseq_contrasts(ddsx, compr, alpha = 0.05, shrink = TRUE)
}
\arguments{
\item{ddsx}{A DESeq object, output from \code{\link[=deseq_wide]{deseq_wide()}}.}

\item{compr}{A character vector for pairwise comparisons of interest, eg.,
\code{c("4", "1")} from draw 4 with samples from draw 1.}

\item{alpha}{A number. The significance level.}

\item{shrink}{A flag. \code{DESeq2::lfcShrink()} the result? Default yes.}
}
\value{
A dataframe.
}
\description{
Obtain pairwise contrasts with shrunken log2FC from a DESeq  object.
}
\examples{
rs_genes <- arrow::read_feather(
  system.file("extdata", "rs_genes_draw_3_4.feather", package = "mirmisc")
)
rs_meta <- arrow::read_feather(
  system.file("extdata", "rs_meta_draw_3_4.feather", package = "mirmisc")
)
ddsx <- deseq_wide(rs_genes, rs_meta, compvar = "meta_draw")
deseq_contrasts(ddsx, compr = c("4", "3"))

}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
