% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{deseq_pairs}
\alias{deseq_pairs}
\title{Get the \code{\link[mirmodels:deseq]{mirmodels::deseq()}} tables from pairs of conditions.}
\usage{
deseq_pairs(prepped_rd_input)
}
\arguments{
\item{prepped_rd_input}{A data frame. The output of a call to
\code{\link[=prep_rd_input]{prep_rd_input()}}.}
}
\value{
A dataframe. A series of \code{\link[mirmodels:deseq]{mirmodels::deseq()}} tables bound together
with \code{\link[dplyr:bind]{dplyr::bind_rows()}}. There are two extra columns \code{cond1} and \code{cond2}
to record the conditions being compared.
}
\description{
Technical replicates in \code{genes_counts} should be collapsed with
\code{\link[=collapse_tech_reps]{collapse_tech_reps()}} prior to running this function.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
