% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{deseq_wide}
\alias{deseq_wide}
\title{Obtain DESeq object from raw counts wide dataframe.}
\usage{
deseq_wide(data, design, compvar)
}
\arguments{
\item{data}{A wide datafame with raw counts that contains raw counts for
samples of interest (may contain other samples as well, which will be
filtered out).}

\item{design}{A dataframe with all of the variables specified in \code{compvar} as
column names. Must have the same number of rows as \code{data} with samples in
the same order.}

\item{compvar}{\itemize{
\item a character vector with the name(s) of the variable(s),
which must be identical to the variable name(s) in the design dataframe.
}}
}
\value{
A DESeq2 object.
}
\description{
Wide is samples in columns, genes in rows.
}
\examples{
rs_genes <- arrow::read_feather(
  system.file("extdata", "rs_genes_draw_3_4.feather", package = "mirmisc")
)
rs_meta <- arrow::read_feather(
  system.file("extdata", "rs_meta_draw_3_4.feather", package = "mirmisc")
)
deseq_wide(rs_genes, rs_meta, compvar = "meta_draw")

}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
