% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{df_fill_missing_genes}
\alias{df_fill_missing_genes}
\title{Fill missing gene columns in one data frame with those from another.}
\usage{
df_fill_missing_genes(df, df_fill_from)
}
\arguments{
\item{df}{A data frame.}

\item{df_fill_from}{A data frame with the same number of rows as \code{df}.}
}
\value{
A data frame.
}
\description{
The 'gene names' are those returned by \code{\link[=get_gene_names]{get_gene_names()}}. This function
takes two data frames \code{df} and \code{df_fill_from} and, if there are any columns
in \code{df_fill_from} with gene names as column names which don't exist in \code{df},
they are copied into \code{df}.
}
\details{
If the gene names in \code{df} are contiguously located, the copied genes are
inserted right after those. Otherwise, they are inserted on the end.
}
\examples{
if (require("mirmodels")) {
  st_data <- get_st_data()
  st_data_median0 <- get_st_data(gene_predicate = ~ median(.) == 0)
  dim(st_data)
  dim(st_data_median0)
  dim(df_fill_missing_genes(st_data_median0, st_data))
}
}
