% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{fgsea_basic}
\alias{fgsea_basic}
\title{Get results from fast Gene Set Enrichment Analysis (GSEA)}
\usage{
fgsea_basic(prepped_gsea_input_lfc, prepped_gsea_input_gene_set)
}
\arguments{
\item{prepped_gsea_input_lfc}{Output of \code{\link[=prep_gsea_input_lfc]{prep_gsea_input_lfc()}}.}

\item{prepped_gsea_input_gene_set}{Output of \code{\link[=prep_gsea_input_gene_set]{prep_gsea_input_gene_set()}}.}
}
\value{
A dataframe.
}
\description{
Uses \code{\link[fgsea:fgsea]{fgsea::fgsea()}} defaults.
}
\examples{
rs_genes <- arrow::read_feather(
  system.file("extdata", "rs_genes_draw_3_4.feather", package = "mirmisc")
)
rs_meta <- arrow::read_feather(
  system.file("extdata", "rs_meta_draw_3_4.feather", package = "mirmisc")
)
ddsx <- deseq_wide(rs_genes, rs_meta, compvar = "meta_draw")
dsq_contrasts <- deseq_contrasts(ddsx, compr = c("4", "3"))
prepped_gsea_input_lfc <- prep_gsea_input_lfc(dsq_contrasts)
prepped_gsea_input_gene_set <- prep_gsea_input_gene_set("C8")
fgsea_basic(prepped_gsea_input_lfc, prepped_gsea_input_gene_set)

}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
