% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes using a predicate function.}
\usage{
filter_genes(df, f)
}
\arguments{
\item{df}{A data frame some of whose column names are gene names.}

\item{f}{A predicate function or a formula coercible to a function by
\code{\link[rlang:as_function]{rlang::as_function()}}.}
}
\value{
A data frame.
}
\description{
Gene names are elements of \code{\link[=get_gene_names]{get_gene_names()}}. Genes for whom the predicate
function \code{f()} evaluates to \code{FALSE} are dropped.
}
