% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{get_df_gene_names}
\alias{get_df_gene_names}
\title{Which gene names are also column names?}
\usage{
get_df_gene_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A character vector.
}
\description{
This is just \code{intersect(names(df), get_gene_names())}.
}
