% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_feather_path}
\alias{get_feather_path}
\title{Get the path to the folder containing the feather files.}
\usage{
get_feather_path(..., use_dotenv = TRUE, verify = TRUE)
}
\arguments{
\item{...}{Character vectors. Elements of the path. Mostly, you'll leave this
blank.}

\item{use_dotenv}{A flag. If the \code{MIRVIE_FEATHER_PATH} environment variable
isn't found by the usual R means, check the \verb{~/.env} file used by
\code{python-dotenv}.}

\item{verify}{A flag. Check that \code{MIRVIE_FEATHER_PATH} exists and contains at
least one *.feather file. Error if check fails. Default \code{TRUE}.}
}
\value{
An \link[fs:path]{fs::path}.
}
\description{
This function requires you to have set the environment variable
\code{MIRVIE_FEATHER_PATH}, which you can do in the \verb{~/.Rprofile} file. It should
have a line like \code{Sys.setenv(MIRVIE_FEATHER_PATH = "path/to/mirvie/feathers/dir")}. If the file doesn't exist, create it and
make this the only line in the file. If this is done correctly, this function
then forms a path with \code{MIRVIE_FEATHER_PATH} as the root directory.
}
\details{
There's a whole vignette explaining this function. To find it, run
\code{browseVignettes(package = "mirmisc")}.
}
\examples{
\dontrun{
get_feather_path()
}

}
