% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{get_gene_names}
\alias{get_gene_names}
\title{Get the names of all of the genes that we use.}
\usage{
get_gene_names()
}
\value{
A character vector.
}
\description{
This includes the Ensembl and colloquial names (so in that sense there is
duplication).
}
\examples{
get_gene_names()
}
