% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirviemisc-package.R
\docType{package}
\name{mirmisc-package}
\alias{mirmisc}
\alias{mirmisc-package}
\title{mirmisc: Non-Modeling Helper Functions For Mirvie R Coders}
\description{
We have random bits of code that are quite useful. This is a home for it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/Mirvie/mirmisc}
  \item Report bugs at \url{https://gitlab.com/Mirvie/mirmisc/-/issues}
}

}
\author{
\strong{Maintainer}: Rory Nolan \email{rorynolan@mirvie.com}

Authors:
\itemize{
  \item Arkady Khodursky \email{arkady@mirvie.com}
  \item Farooq Siddiqui \email{farooq@mirvie.com}
  \item Elaine Gee \email{elaine@mirvie.com}
  \item Joan Camunas Soler \email{joan@mirvie.com}
  \item Morten Rasmussen \email{morten@mirvie.com}
}

}
\keyword{internal}
