% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-names.R
\name{mutate_genes}
\alias{mutate_genes}
\title{Apply the same function to all columns whose names are gene names.}
\usage{
mutate_genes(df, f)
}
\arguments{
\item{df}{A data frame.}

\item{f}{A function or a formula coercible to a function by
\code{\link[rlang:as_function]{rlang::as_function()}}.}
}
\value{
A data frame.
}
\description{
Gene names are elements of \code{\link[=get_gene_names]{get_gene_names()}}.
}
