% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{plot_cond_paired_pearson}
\alias{plot_cond_paired_pearson}
\title{Violin plots for within-condition gene Pearson correlation coefficients.}
\usage{
plot_cond_paired_pearson(prepped_rd_input, gene_subset = NULL, qntl = NULL)
}
\arguments{
\item{prepped_rd_input}{A data frame. The output of a call to
\code{\link[=prep_rd_input]{prep_rd_input()}}.}

\item{gene_subset}{A character vector of genes in a subset of interest.}

\item{qntl}{A number between 0 and 1. A quantile-based threshold below which
the data is subset. The purpose of this is to improve the sensitivity of
the reproducibility analysis by focusing on subsets of data with lower
counts.}
}
\value{
A ggplot.
}
\description{
Violin plots for within-condition gene Pearson correlation coefficients.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
