% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{plot_down}
\alias{plot_down}
\alias{plot_up}
\title{Plot top \code{n} most significantly enriched pathways in the upward or downward
direction.}
\usage{
plot_down(fgsea_out, padj_thresh = 0.05, n = 20)

plot_up(fgsea_out, padj_thresh = 0.05, n = 20)
}
\arguments{
\item{fgsea_out}{Output of \code{\link[fgsea:fgsea]{fgsea::fgsea()}} or \code{\link[=fgsea_basic]{fgsea_basic()}}.}

\item{padj_thresh}{Adjusted p-value threshold.}

\item{size_thresh}{number of genes to show.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Plot top \code{n} most significantly enriched pathways in the upward or downward
direction.
}
\examples{
rs_genes <- arrow::read_feather(
  system.file("extdata", "rs_genes_draw_3_4.feather", package = "mirmisc")
)
rs_meta <- arrow::read_feather(
  system.file("extdata", "rs_meta_draw_3_4.feather", package = "mirmisc")
)
ddsx <- deseq_wide(rs_genes, rs_meta, compvar = "meta_draw")
dsq_contrasts <- deseq_contrasts(ddsx, compr = c("4", "3"))
prepped_gsea_input_lfc <- prep_gsea_input_lfc(dsq_contrasts)
prepped_gsea_input_gene_set <- prep_gsea_input_gene_set("C8")
fgs <- fgsea_basic(prepped_gsea_input_lfc, prepped_gsea_input_gene_set)
plot_down(fgs)
plot_up(fgs)

}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
