% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{plot_metrics_global}
\alias{plot_metrics_global}
\title{Plots comparisons between conditions in an experiment.}
\usage{
plot_metrics_global(prepped_rd_input, gene_subset = NULL)
}
\arguments{
\item{prepped_rd_input}{A data frame. The output of a call to
\code{\link[=prep_rd_input]{prep_rd_input()}}.}

\item{gene_subset}{A character vector of genes in a subset of interest.}
}
\value{
A list of 6 ggplots.
}
\description{
Actual experimental samples and positive controls, using all genes in the
count table or a subset of genes.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
