% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{plot_var_mean_ratio}
\alias{plot_var_mean_ratio}
\title{Plot median and other quantile values for variance-to-mean ratios obtained
across all genes.}
\usage{
plot_var_mean_ratio(prepped_rd_input)
}
\arguments{
\item{prepped_rd_input}{A data frame. The output of a call to
\code{\link[=prep_rd_input]{prep_rd_input()}}.}
}
\value{
A ggplot.
}
\description{
Useful in comparison to a reference: if median and other quantile ratios are
higher or lower than in the reference, then one can draw a conclusion about a
relative variability among the samples in a given experimental condition.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
