% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{prep_gsea_input_gene_set}
\alias{prep_gsea_input_gene_set}
\title{Prepare gene set input for \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.}
\usage{
prep_gsea_input_gene_set(category)
}
\arguments{
\item{cetegory}{A string. Category of interest.}
}
\value{
A list object for fgsea.
}
\description{
Categories of interest from
http://www.gsea-msigdb.org/gsea/msigdb/index.jsp.
}
\examples{
prep_gsea_input_gene_set("C8")

}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{prep_rd_input}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
