% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkady.R
\name{prep_rd_input}
\alias{prep_rd_input}
\title{Join the sample descriptions and genes counts data frames into a wide format.}
\usage{
prep_rd_input(sample_descs, genes_counts)
}
\arguments{
\item{sample_descs}{A data frame of sample descriptions. The \code{sampleName}
column holds the sample names.}

\item{genes_counts}{A data frame of gene counts. The first column is \code{gene}
and the rest are sample names.}
}
\value{
A data frame. The gene counts table with columns corresponding to the
desired subset of data.
}
\description{
This function calls \code{\link[=reconcile_names]{reconcile_names()}} and \code{\link[=check_sample_descs]{check_sample_descs()}}
internally.
}
\seealso{
Other Arkady: 
\code{\link{check_sample_descs}()},
\code{\link{collapse_tech_reps}()},
\code{\link{deseq_contrasts}()},
\code{\link{deseq_pairs}()},
\code{\link{deseq_wide}()},
\code{\link{fgsea_basic}()},
\code{\link{plot_cond_paired_pearson}()},
\code{\link{plot_down}()},
\code{\link{plot_metrics_global}()},
\code{\link{plot_var_mean_ratio}()},
\code{\link{prep_gsea_input_gene_set}()},
\code{\link{prep_gsea_input_lfc}()},
\code{\link{reconcile_names}()}
}
\concept{Arkady}
