% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsim.R
\name{repsim_gene_cond}
\alias{repsim_gene_cond}
\title{Repeatedly simulate gene counts for cases and controls of a condition.}
\usage{
repsim_gene_cond(
  n_gene_reps,
  n_samples,
  cond_prevalence,
  control_mean,
  case_mean
)
}
\arguments{
\item{n_gene_reps}{The number of times to repeatedly simulate the gene counts
for that gene.}

\item{n_samples}{The number of samples in the simulation.}

\item{cond_prevalence}{A number in (0, 1). The condition prevalence.}

\item{control_mean}{Theoretical mean of the gene for controls.}

\item{case_mean}{Theoretical mean of the gene for cases.}
}
\value{
A tibble with \code{n_gene_reps + 1} columns called \code{generep_1},
\code{generep_2}, . . ., \verb{generep_\{n_gene_reps\}}, \code{cond} and \code{n_samples}
columns.
}
\description{
Given a number of gene repetitions, a number of samples, the condition
prevalence and the theoretical mean of a gene for cases and controls,
repeatedly simulate gene counts for that gene across the condition.
}
\details{
Poisson gene counts are assumed.
}
\examples{
if (rlang::is_installed("mirmodels")) {
  rsgc5000 <- repsim_gene_cond(15000, 5000, 1 / 10, 0.01, 0.05)
  sde <- mirmodels::cor_de(rsgc5000, "cond", head(names(rsgc5000), -1))
}
}
